# doc/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AriesHub API",
    "version": "0.7.0",
    "description": "Public content catalog, Redis-backed cookie identity, flat paged discussions with comment likes, administrator content and moderation API. Credit balance lives on users and pricing on publications; the credit ledger and unlock tables exist but the credit application API is not implemented yet. Private S3-compatible file uploads and owner-only temporary downloads. All JSON responses use Result {code,msg,data,traceId}; success code is SUCCESS, failures retain semantic HTTP status and business codes. Empty successes use HTTP 200 with data:null; throttling includes Retry-After and data.retryAfterSeconds."
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1"
    }
  ],
  "paths": {
...
    "/admin/categories": {
      "get": {
        "operationId": "getAdminCategories",
        "summary": "Lists category options for the content editor",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Category options",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdminCategory"
                      }
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/categories": {
      "get": {
        "operationId": "getCategories",
        "summary": "Categories with published and available publication counts",
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Category"
                      }
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database temporarily unavailable",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/publications": {
      "get": {
        "operationId": "getAdminPublications",
        "summary": "Lists publications in every publication state",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All publications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdminPublicationSummary"
                      }
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "operationId": "createAdminPublication",
        "summary": "Creates a draft publication and its content",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminPublicationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Draft created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AdminPublicationDetail"
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/admin/publications/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/PublicationId"
        }
      ],
      "get": {
        "operationId": "getAdminPublication",
        "summary": "Returns the complete editable publication",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Editable publication",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AdminPublicationDetail"
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "operationId": "updateAdminPublication",
        "summary": "Replaces the editable publication fields and content",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminPublicationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Publication updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AdminPublicationDetail"
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/admin/publications/{id}/publish": {
      "post": {
        "operationId": "publishAdminPublication",
        "summary": "Publishes a publication to the public catalog",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PublicationId"
          }
        ],
        "responses": {
          "200": {
            "description": "Published publication",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AdminPublicationDetail"
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/admin/publications/{id}/archive": {
      "post": {
        "operationId": "archiveAdminPublication",
        "summary": "Archives a publication and removes it from the public catalog",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PublicationId"
          }
        ],
        "responses": {
          "200": {
            "description": "Archived publication",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AdminPublicationDetail"
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/publications": {
      "get": {
        "operationId": "getPublications",
        "summary": "Public publication summaries; never includes body content",
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/PublicationPage"
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database temporarily unavailable",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10000,
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 24,
              "default": 9
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 120,
              "description": "Trimmed literal substring matched against title/summary, publication insensitive. SQL wildcards are escaped."
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9-]{0,80}$"
            }
          },
          {
            "name": "access",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "",
                "FREE",
                "CREDIT"
              ]
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "",
                "CASE_STUDY",
                "ARTICLE",
                "COURSE"
              ]
            }
          }
        ]
      }
    },
    "/publications/{slug}": {
      "get": {
        "operationId": "getPublication",
        "summary": "Public detail and preview; never includes full body",
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/PublicationDetail"
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid slug",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing, draft, archived or suspended publication",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database temporarily unavailable",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              },
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9-]{1,120}$"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Category": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "slug",
          "name",
          "publicationCount"
        ],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9]+$",
            "description": "PostgreSQL bigint ID serialized as a string."
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "publicationCount": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "AdminCategory": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "slug",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "msg",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "retryAfterSeconds"
                ],
                "properties": {
                  "retryAfterSeconds": {
                    "type": "integer",
                    "minimum": 1
                  }
                }
              }
            ]
          },
          "traceId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    },
...
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Not signed in",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Signed in without the ADMIN role",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "Publication slug already exists",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "arieshub_token"
      }
    }
  }
}